update context switch penalty, minor refactoring variable names, dynamically allocating remaining time to appease the CSE machines

+++ src/ProcessSimulator.h
@@ -66,8 +66,8 @@ private:
 	unsigned int waitingTime;
 	unsigned int totalRunningTime;
 	SchedulingAlgorithim currentAlgorithim;
-	unsigned int contextPenalty;
-	unsigned int totalPenalty;
+	unsigned long contextPenalty;
+	unsigned long totalPenalty;
 	ProcessSet scheduledProcesses;
 	unsigned int numberProcesses;
 	bool multiProcessMode;
